Move authentication errors to be accessible in the index scheduler
