+++ crates/meilisearch/src/routes/multi_search.rs
@@ -4,7 +4,7 @@ use actix_web::{HttpRequest, HttpResponse};
 use deserr::actix_web::AwebJson;
 use index_scheduler::IndexScheduler;
 use meilisearch_types::deserr::DeserrJsonError;
-use meilisearch_types::error::{Code, ResponseError};
+use meilisearch_types::error::{AuthenticationError, Code, ResponseError};
 use meilisearch_types::keys::actions;
 use meilisearch_types::milli::progress::Progress;
 use meilisearch_types::milli::TotalProcessingTimeStep;
@@ -18,7 +18,7 @@ use crate::analytics::Analytics;
 use crate::documents_retrieval::{preprocess_filters, DocumentSearch, DocumentSearchResult};
 use crate::error::MeilisearchHttpError;
 use crate::extractors::authentication::policies::ActionPolicy;
-use crate::extractors::authentication::{AuthenticationError, GuardedData};
+use crate::extractors::authentication::GuardedData;
 use crate::personalization::PersonalizationService;
 use crate::routes::parse_include_metadata_header;
 use crate::search::federated::types::PreprocessedQuery;
